+++ core/SConscript.bootloader
@@ -94,6 +94,7 @@ SOURCE_BOOTLOADER = [
 
 SOURCE_TREZORHAL = [
     'embed/trezorhal/common.c',
+    'embed/trezorhal/drbg.c',
     'embed/trezorhal/entropy.c',
     'embed/trezorhal/image.c',
     'embed/trezorhal/flash.c',
