+++ core/SConscript.firmware
@@ -330,6 +330,7 @@ SOURCE_TREZORHAL = [
     'embed/trezorhal/common.c',
     'embed/trezorhal/dma.c',
     'embed/trezorhal/entropy.c',
+    'embed/trezorhal/drbg.c',
     'embed/trezorhal/flash.c',
     'embed/trezorhal/mini_printf.c',
     'embed/trezorhal/mpu.c',
