+++ core/embed/bootloader/main.c
@@ -22,6 +22,7 @@
 
 #include "common.h"
 #include "display.h"
+#include "drbg.h"
 #include "flash.h"
 #include "image.h"
 #include "mini_printf.h"
