+++ core/embed/extmod/modtrezorconfig/modtrezorconfig.c
@@ -27,6 +27,7 @@
 
 #include "embed/extmod/trezorobj.h"
 
+#include "drbg.h"
 #include "entropy.h"
 #include "memzero.h"
 #include "storage.h"
