+++ core/embed/trezorhal/common.c
@@ -23,13 +23,9 @@
 
 #include "common.h"
 #include "display.h"
-#include "hmac_drbg.h"
-#include "rand.h"
 
 // from util.s
 extern void shutdown(void);
-
-static HMAC_DRBG_CTX drbg_ctx;
 
 #define COLOR_FATAL_ERROR RGB16(0x7F, 0x00, 0x00)
 
@@ -140,24 +136,4 @@ uint32_t __stack_chk_guard = 0;
 
 void __attribute__((noreturn)) __stack_chk_fail(void) {
   error_shutdown("Internal error", "(SS)", NULL, NULL);
-}
-
-void drbg_init(void) {
-  uint8_t entropy[48];
-  random_buffer(entropy, sizeof(entropy));
-  hmac_drbg_init(&drbg_ctx, entropy, sizeof(entropy), NULL, 0);
-}
-
-void drbg_reseed(const uint8_t *entropy, size_t len) {
-  hmac_drbg_reseed(&drbg_ctx, entropy, len, NULL, 0);
-}
-
-void drbg_generate(uint8_t *buf, size_t len) {
-  hmac_drbg_generate(&drbg_ctx, buf, len);
-}
-
-uint32_t drbg_random32(void) {
-  uint32_t value;
-  drbg_generate((uint8_t *)&value, sizeof(value));
-  return value;
 }
